Refactored two methods from the UserController
They now live in the BookController given the responsibility of both methods. They don't really seem to belong in the User Controller. See comments for more.
